% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{adtestls1}
\alias{adtestls1}
\title{Helper Function for adtestl Bootstrap}
\usage{
adtestls1(vstar, yhat, res, mflag, x, fr)
}
\arguments{
\item{vstar}{Bootstrap multiplier vector.}

\item{yhat}{Fitted values from additive model.}

\item{res}{Residuals from additive model.}

\item{mflag}{Matrix indicating coordinate-wise ordering.}

\item{x}{Matrix of predictor variables.}

\item{fr}{Span parameter for loess smoother.}
}
\value{
Vector of residual process values.
}
\description{
Internal helper function used by \code{\link{adtestl}} to compute bootstrap
residual process values using loess smoother.
}
\keyword{internal}
