% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{adtests1}
\alias{adtests1}
\title{Helper Function for adtest Bootstrap}
\usage{
adtests1(vstar, yhat, res, mflag, x, fr)
}
\arguments{
\item{vstar}{Bootstrap multiplier vector.}

\item{yhat}{Fitted values from additive model.}

\item{res}{Residuals from additive model.}

\item{mflag}{Matrix indicating coordinate-wise ordering.}

\item{x}{Matrix of predictor variables.}

\item{fr}{Span parameter for running interval smoother.}
}
\value{
Vector of residual process values.
}
\description{
Internal helper function used by \code{\link{adtest}} to compute bootstrap
residual process values using running interval smoother.
}
\keyword{internal}
