% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{anc.best.ex}
\alias{anc.best.ex}
\title{Extract P-Values for Best Group Comparisons}
\usage{
anc.best.ex(x, tr = 0.2)
}
\arguments{
\item{x}{Matrix or list of data where each column (or list element) is a group.}

\item{tr}{Trimming proportion (default: 0.2). Proportion of observations to
trim from each tail when computing winsorized correlation.}
}
\value{
Vector of p-values from comparing best group to each other group.
}
\description{
Helper function for anc.best.crit. Computes p-values comparing the group with
the largest trimmed mean to all other groups.
}
\keyword{internal}
