% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{anclinQS.plot}
\alias{anclinQS.plot}
\title{Plot Quantile Shift Effect Sizes}
\usage{
anclinQS.plot(
  x1,
  y1,
  x2,
  y2,
  pts = NULL,
  q = 0.1,
  xout = FALSE,
  ALL = TRUE,
  npts = 10,
  line = TRUE,
  xlab = "X",
  ylab = "QS.Effect",
  outfun = outpro,
  REQMIN = 0.001,
  ...
)
}
\arguments{
\item{x1}{Covariate values for control group (single covariate).}

\item{y1}{Outcome values for control group.}

\item{x2}{Covariate values for experimental group.}

\item{y2}{Outcome values for experimental group.}

\item{pts}{Covariate points where effect sizes are plotted. If NULL, chosen automatically.}

\item{q}{Quantile for determining covariate range. Default is 0.1.}

\item{xout}{Logical. Remove outliers from covariates. Default is FALSE.}

\item{ALL}{Logical. If TRUE, use many points. Default is TRUE.}

\item{npts}{Number of points if pts=NULL. Default is 10.}

\item{line}{Logical. Connect points with lines. Default is TRUE.}

\item{xlab, ylab}{Axis labels.}

\item{outfun}{Outlier detection function. Default is outpro.}

\item{REQMIN}{Minimum required value for qinvreg. Default is 0.001.}

\item{...}{Additional arguments.}
}
\value{
Matrix with columns: Pts, Y.hat4ExpGrp, QS.Effect.Size (returned invisibly).
}
\description{
Plot quantile shift (QS) effect sizes as a function of covariate values for two independent groups.
The QS effect size is the conditional quantile of the control group corresponding to the median
of Y for the experimental group, given X.
}
\keyword{ancova}
\keyword{effect-size}
\keyword{plot}
