% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{apanova}
\alias{apanova}
\title{Agresti-Pendergast Rank Test for Dependent Groups}
\usage{
apanova(data, grp = 0)
}
\arguments{
\item{data}{Data in matrix form (n by J) or in list mode where each element
contains data for one group. For dependent groups, rows represent matched
observations.}

\item{grp}{A vector indicating group membership. Can be numeric, factor, or
character. Used for grouping observations in multi-group comparisons.}
}
\value{
A list with components:
\item{FTEST}{The F test statistic.}
\item{df1}{Numerator degrees of freedom (J-1).}
\item{df2}{Denominator degrees of freedom ((J-1)(n-1)).}
\item{p.value}{The p-value for the test.}
}
\description{
Performs an Agresti-Pendergast rank test for comparing J dependent (repeated
measures) groups. This is a nonparametric alternative to repeated measures ANOVA.
}
\details{
The Agresti-Pendergast test uses ranks to test for differences among J
dependent groups. It is based on an F-statistic with (J-1) and (J-1)(n-1)
degrees of freedom.

For n <= 20, the function suggests using \code{bprm} instead for better
small sample performance.
}
\examples{
# Three dependent measurements
set.seed(123)
time1 <- rnorm(15, mean=10)
time2 <- time1 + rnorm(15, mean=0.5, sd=0.5)
time3 <- time1 + rnorm(15, mean=1, sd=0.5)
apanova(cbind(time1, time2, time3))
}
\seealso{
\code{\link{rmanova}}, \code{\link{bprm}}
}
