% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{bb.es.main}
\alias{bb.es.main}
\title{Main Effect Sizes for Between-by-Between ANOVA}
\usage{
bb.es.main(J, K, x, DIF = TRUE, ...)
}
\arguments{
\item{J}{Number of levels for Factor A (between-subjects)}

\item{K}{Number of levels for Factor B (between-subjects)}

\item{x}{Data in matrix/data frame format (will be converted to list) or
as a list of length J*K}

\item{DIF}{Not used in this function (included for consistency with \code{\link{bw.es.main}})}

\item{...}{Additional arguments passed to \code{\link{t1wayv2}}}
}
\value{
A list with components:
\item{A.explanatory.effect.size}{Effect size for Factor A (between-subjects)}
\item{B.explanatory.effect.size}{Effect size for Factor B (between-subjects)}
}
\description{
Computes main effect sizes for a two-way ANOVA with two between-subjects
factors (A and B).
}
\details{
For both Factor A and Factor B (both between-subjects), computes explanatory
effect sizes using \code{\link{t1wayv2}}. Factor A effect size pools
observations across all levels of Factor B, and Factor B effect size pools
observations across all levels of Factor A.

This differs from \code{\link{bw.es.main}}, which uses projection effect
sizes for the within-subjects factor.
}
\seealso{
\code{\link{bw.es.main}}, \code{\link{t1wayv2}}
}
\keyword{htest}
