% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{bdmP}
\alias{bdmP}
\title{Brunner-Dette-Munk Test for One-Way Design}
\usage{
bdmP(x)
}
\arguments{
\item{x}{Data in list mode (each element is a group) or matrix (columns are groups)}
}
\value{
A list with components:
\item{test.stat}{ANOVA-type test statistic}
\item{df1}{Numerator degrees of freedom}
\item{df2}{Denominator degrees of freedom}
\item{p.value}{p-value for the test}
\item{q.hat}{Estimated relative effects from \code{bdm}}
}
\description{
Tests the null hypothesis from Brunner et al. (2016) for a one-way design
using the rankFD package. Tests for differences in distributions across
independent groups.
}
\details{
This function implements the method from Brunner et al. (2016) for testing
the null hypothesis H0p (purely nonparametric hypothesis). It uses rank-based
methods via the rankFD package.

The function converts list data to a data frame with group indicators and
calls \code{rankFD} with hypothesis='H0p'. It also computes relative effects
using the \code{bdm} function.
}
\references{
Brunner et al. (2016). \link{Complete reference for the BDM method}
}
\seealso{
\code{\link{bdm}}
}
