% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootdepsub}
\alias{bootdepsub}
\title{Bootstrap Helper Function for Dependent Data}
\usage{
bootdepsub(isub, x, tr)
}
\arguments{
\item{isub}{Vector of indices for the bootstrap sample.}

\item{x}{Data vector.}

\item{tr}{Trimming proportion.}
}
\value{
Trimmed mean of the bootstrap sample.
}
\description{
Internal helper function used by \code{bootdep} to compute trimmed means
for bootstrap samples.
}
\keyword{internal}
