% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{bw.es.main}
\alias{bw.es.main}
\title{Main Effect Sizes for Between-by-Within ANOVA}
\usage{
bw.es.main(J, K, x, DIF = TRUE, ...)
}
\arguments{
\item{J}{Number of levels for Factor A (between-subjects)}

\item{K}{Number of levels for Factor B (within-subjects)}

\item{x}{Data in matrix/data frame format (will be converted to list) or
as a list of length J*K}

\item{DIF}{Logical. If TRUE (default), use \code{\link{rmES.dif.pro}} for
Factor B effect size. If FALSE, use \code{\link{rmES.pro}}.}

\item{...}{Additional arguments passed to \code{\link{t1wayv2}} and effect size functions}
}
\value{
A list with components:
\item{A.explanatory.effect.size}{Effect size for Factor A (between-subjects)}
\item{B.projection.effect.size}{Effect size for Factor B (within-subjects)}
}
\description{
Computes main effect sizes for a two-way ANOVA with one between-subjects
factor (A) and one within-subjects factor (B).
}
\details{
For Factor A (between-subjects), computes explanatory effect size using
\code{\link{t1wayv2}} by pooling observations across all levels of Factor B.

For Factor B (within-subjects), computes projection effect size using either
\code{\link{rmES.dif.pro}} (when DIF=TRUE, based on difference scores) or
\code{\link{rmES.pro}} (when DIF=FALSE, based on raw scores), pooling
observations across all levels of Factor A.
}
\seealso{
\code{\link{bb.es.main}}, \code{\link{t1wayv2}},
\code{\link{rmES.pro}}, \code{\link{rmES.dif.pro}}
}
\keyword{htest}
