% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{chi.int.p}
\alias{chi.int.p}
\title{Chi-Squared Integration Helper (Positive)}
\usage{
chi.int.p(p, a, c1)
}
\arguments{
\item{p}{Number of dimensions}

\item{a}{Additional degrees of freedom parameter}

\item{c1}{Integration constant}
}
\value{
Numeric value of the integral expression
}
\description{
Helper function for computing integrals involving chi-squared distributions
in robust estimation contexts. Internal utility function.
}
\keyword{internal}
