% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{chi.int2.p}
\alias{chi.int2.p}
\title{Chi-Squared Integration Helper (Negative)}
\usage{
chi.int2.p(p, a, c1)
}
\arguments{
\item{p}{Number of dimensions}

\item{a}{Additional degrees of freedom parameter}

\item{c1}{Integration constant}
}
\value{
Numeric value of the integral expression (negative)
}
\description{
Helper function for computing integrals involving chi-squared distributions
in robust estimation contexts. Returns negative of \code{chi.int.p}.
}
\seealso{
\code{\link{chi.int.p}}
}
\keyword{internal}
