% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{cohen2xi}
\alias{cohen2xi}
\title{Convert Cohen's d to Xi Effect Size}
\usage{
cohen2xi(delta)
}
\arguments{
\item{delta}{Cohen's d effect size value}
}
\value{
Numeric value of xi effect size
}
\description{
Converts Cohen's d effect size to xi (ξ), a measure used in robust statistics.
}
\details{
Uses the transformation: xi = sqrt((2*delta^2)/(4 + delta^2))

This conversion is useful when working with robust effect size measures.
}
\seealso{
\code{\link{xi2cohen}} for the inverse transformation
}
\keyword{univariate}
