% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-core.R
\name{con3way}
\alias{con3way}
\title{Create Contrast Coefficients for Three-Way ANOVA}
\usage{
con3way(J, K, L)
}
\arguments{
\item{J}{Number of levels for Factor A.}

\item{K}{Number of levels for Factor B.}

\item{L}{Number of levels for Factor C.}
}
\value{
A list with seven components for all effects:
\item{conA, conB, conC}{Main effect contrasts for factors A, B, C}
\item{conAB, conAC, conBC}{Two-way interaction contrasts}
\item{conABC}{Three-way interaction contrasts}
}
\description{
Generates all contrast coefficient matrices for performing pairwise comparisons
of main effects and interactions in a J x K x L three-way ANOVA design.
}
\details{
For a J x K x L factorial ANOVA, this function generates all necessary
contrast matrices for testing main effects and all possible interactions.
}
\examples{
# 2x2x2 design
con3way(2, 2, 2)

}
\seealso{
\code{\link{con2way}} for two-way designs
}
