% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{conc2}
\alias{conc2}
\title{Find LTA Attractor Using Concentration Algorithm}
\usage{
conc2(x, y, start = l1fit(x, y)$coef)
}
\arguments{
\item{x}{Design matrix (without intercept column)}

\item{y}{Response vector}

\item{start}{Starting coefficient vector (default: L1 regression fit)}
}
\value{
List with components:
\item{coef}{Final coefficient vector (intercept and slopes)}
\item{ct}{Number of concentration steps performed}
}
\description{
Finds the LTA (Least Trimmed sum of Absolute deviations) attractor starting
from a given initial coefficient vector using a concentration algorithm.
}
\details{
Iteratively refits L1 regression to observations with smallest absolute residuals
(h = ceiling(n/2) observations) until convergence of the LTA objective function.
}
\seealso{
\code{\link{l1fit}}, \code{\link{ltareg}}
}
\keyword{regression}
\keyword{robust}
