% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{cor2xy}
\alias{cor2xy}
\title{Correlate Two Predictors with a Response Variable}
\usage{
cor2xy(x, y, corfun = spear, ...)
}
\arguments{
\item{x}{Matrix with two columns containing the predictor variables}

\item{y}{Response variable (numeric vector)}

\item{corfun}{Correlation function to use (default: spear)}

\item{...}{Additional arguments passed to corfun}
}
\value{
A list with component:
\item{cor}{Vector of length 2 containing the two correlation coefficients}
}
\description{
Computes the correlation between each of two predictor variables and a
response variable using a specified correlation function.
}
\details{
This function computes the correlation between x\link{,1} and y, and the correlation
between x\link{,2} and y, using the specified correlation function. Any correlation
function that returns a list with a 'cor' component can be used.
}
\seealso{
\code{\link{spear}}, \code{\link{COR.PAIR}}
}
\keyword{multivariate}
