% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov2cor}
\alias{cov2cor}
\title{Convert Covariance Matrix to Correlation Matrix}
\usage{
cov2cor(x)
}
\arguments{
\item{x}{A p × p covariance matrix}
}
\value{
A p × p correlation matrix
}
\description{
Transforms a covariance matrix into the corresponding correlation matrix by
standardizing.
}
\details{
Computes correlations as \code{cor[i,j] = cov[i,j] / sqrt(cov[i,i] * cov[j,j])}.

Note: Base R also has \code{cov2cor} which accomplishes the same task. This
function provides an alternative implementation within the WRS package.
}
\examples{
\dontrun{
x <- matrix(rnorm(300), ncol = 3)
C <- cov(x)
R <- cov2cor(C)  # Convert to correlation
all.equal(diag(R), rep(1, 3))  # Diagonal should be 1
}
}
\seealso{
Other covariance estimation: 
\code{\link{DETMCD}()},
\code{\link{Scov}()},
\code{\link{bbwcovm}()},
\code{\link{bicov}()},
\code{\link{bicovM}()},
\code{\link{bicovm}()},
\code{\link{bwwcovm}()},
\code{\link{cov.mba}()},
\code{\link{cov.ogk}()},
\code{\link{cov.roc}()},
\code{\link{cov2med}()},
\code{\link{covl}()},
\code{\link{covmba}()},
\code{\link{covmba2}()},
\code{\link{covmcd}()},
\code{\link{covmmed}()},
\code{\link{covmve}()},
\code{\link{covogk}()},
\code{\link{covout}()},
\code{\link{covroc}()},
\code{\link{dcov}()},
\code{\link{gkcov}()},
\code{\link{longcov2mat}()},
\code{\link{mcdcov}()},
\code{\link{mgvcov}()},
\code{\link{mscov}()},
\code{\link{mvecov}()},
\code{\link{rmba}()},
\code{\link{skip.cov}()},
\code{\link{skipcov}()},
\code{\link{skipogk}()},
\code{\link{tbscov}()},
\code{\link{wincov}()},
\code{\link{wincovN}()},
\code{\link{wmean.cov}()}
}
\concept{covariance estimation}
