% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covmba2}
\alias{covmba2}
\title{Compute MBA Covariance Matrix (Alternative Implementation)}
\usage{
covmba2(x, csteps = 5)
}
\arguments{
\item{x}{Numeric matrix (n × p) of observations}

\item{csteps}{Number of concentration steps (default: 5)}
}
\value{
A list with two components:
\item{center}{p-dimensional robust location estimate}
\item{cov}{p × p robust scatter matrix}
}
\description{
Alternative implementation of the Median Ball Algorithm covariance estimator,
used internally by \code{\link{cov.mba}}.
}
\details{
This function implements the basic MBA algorithm similar to \code{\link{covmba}}.
See \code{\link{covmba}} for algorithm details. For most applications,
\code{\link{rmba}} (reweighted MBA) is preferred for better efficiency.
}
\examples{
\dontrun{
x <- matrix(rnorm(300), ncol = 3)
result <- covmba2(x)
result$center
result$cov
}
}
\seealso{
\code{\link{covmba}} for MBA with both starts,
\code{\link{rmba}} for reweighted MBA (preferred),
\code{\link{cov.mba}} for wrapper extracting covariance only

Other covariance estimation: 
\code{\link{DETMCD}()},
\code{\link{Scov}()},
\code{\link{bbwcovm}()},
\code{\link{bicov}()},
\code{\link{bicovM}()},
\code{\link{bicovm}()},
\code{\link{bwwcovm}()},
\code{\link{cov.mba}()},
\code{\link{cov.ogk}()},
\code{\link{cov.roc}()},
\code{\link{cov2cor}()},
\code{\link{cov2med}()},
\code{\link{covl}()},
\code{\link{covmba}()},
\code{\link{covmcd}()},
\code{\link{covmmed}()},
\code{\link{covmve}()},
\code{\link{covogk}()},
\code{\link{covout}()},
\code{\link{covroc}()},
\code{\link{dcov}()},
\code{\link{gkcov}()},
\code{\link{longcov2mat}()},
\code{\link{mcdcov}()},
\code{\link{mgvcov}()},
\code{\link{mscov}()},
\code{\link{mvecov}()},
\code{\link{rmba}()},
\code{\link{skip.cov}()},
\code{\link{skipcov}()},
\code{\link{skipogk}()},
\code{\link{tbscov}()},
\code{\link{wincov}()},
\code{\link{wincovN}()},
\code{\link{wmean.cov}()}
}
\concept{covariance estimation}
