% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{dep.loc.summary}
\alias{dep.loc.summary}
\title{Summary of Location Estimates for Dependent Groups}
\usage{
dep.loc.summary(x, y)
}
\arguments{
\item{x}{Numeric vector or matrix. For bivariate functions, the first variable.
For matrix functions, an n-by-p data matrix.}

\item{y}{Numeric vector. The second variable for bivariate correlations.
Can be \code{NULL} for some functions that accept a two-column matrix as \code{x}.}
}
\value{
Matrix with various location estimates (mean, trimmed mean, median, M-estimator, Mdif) for each variable.
}
\description{
Computes multiple location estimates for X, Y, and their difference X-Y.
}
\seealso{
\code{\link{loc.dif.summary}}, \code{\link{loc2dif}}
}
