% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{dnormvar}
\alias{dnormvar}
\title{Normal Distribution Variance Helper}
\usage{
dnormvar(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of transformed values.
}
\description{
Internal helper function that computes x^2 * dnorm(x).
}
\keyword{internal}
