% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-core.R
\name{erho.bt}
\alias{erho.bt}
\title{Expectation of Biweight Rho Function}
\usage{
erho.bt(p, c1, M)
}
\arguments{
\item{p}{Degrees of freedom.}

\item{c1}{Tuning constant.}

\item{M}{Location parameter.}
}
\value{
Expected value.
}
\description{
Internal helper function computing expectation of rho function under
chi-squared distribution.
}
\keyword{internal}
