% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{FactorAnalysis}
\alias{FactorAnalysis}
\title{Principal Axis Factor Analysis}
\usage{
FactorAnalysis(
  data,
  corr.matrix = FALSE,
  max.iteration = 50,
  n.factors = 0,
  corr.type = "pearson"
)
}
\arguments{
\item{data}{Matrix of observed data, or correlation matrix if \code{corr.matrix=TRUE}}

\item{corr.matrix}{Logical: if TRUE, \code{data} is a correlation matrix (default: FALSE)}

\item{max.iteration}{Maximum number of iterations for communality estimation (default: 50)}

\item{n.factors}{Number of factors to extract. If 0, determined by eigenvalues > 1 rule (default: 0)}

\item{corr.type}{Type of correlation: "pearson" or "spearman" (default: "pearson")}
}
\value{
List with components:
\item{loadings}{Factor loadings (vector if one factor, matrix if multiple factors)}
\item{factors}{Number of factors extracted}
}
\description{
Performs principal axis factor analysis on data or a correlation matrix.
Extracts factor loadings using an iterative communality estimation procedure.
}
\details{
This function implements principal axis factoring, an iterative procedure that:
\enumerate{
\item Estimates communalities (h²) from the correlation matrix
\item Replaces diagonal of correlation matrix with h² estimates
\item Extracts eigenvalues and eigenvectors
\item Computes factor loadings
\item Iterates until convergence (change < 0.001) or max iterations
}

If \code{n.factors=0}, the number of factors is determined by the eigenvalue > 1 rule.
}
\references{
Ruscio, J. & Kaczetow, W. (2008). Simulating Multivariate Nonnormal Data Using
an Iterative Algorithm. Multivariate Behavioral Research, 43, 355-381.
}
\seealso{
\code{\link{gengh}}
}
\keyword{multivariate}
