% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{ftrim}
\alias{ftrim}
\title{Transformation Function for Trimmed Mean}
\usage{
ftrim(z, tr, g, h)
}
\arguments{
\item{z}{A numeric value or vector.}

\item{tr}{Trimming proportion.}

\item{g, h}{Parameters for g-and-h transformation.}
}
\value{
Transformed value(s).
}
\description{
Internal helper function used in trimmed mean calculations with g-and-h
distributions.
}
\keyword{internal}
