% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{ghdist}
\alias{ghdist}
\title{Generate Random Samples from g-and-h Distribution}
\usage{
ghdist(n, g = 0, h = 0)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{g}{Shape parameter (default: 0).}

\item{h}{Tail weight parameter (default: 0).}
}
\value{
Vector of n random observations from the g-and-h distribution.
}
\description{
Generates random observations from a g-and-h distribution.
}
\details{
The g-and-h distribution is a flexible family for modeling skewness (g) and tail weight (h).
When g = h = 0, reduces to standard normal.
}
\seealso{
\code{\link{ghmean}}
}
