% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{l2v}
\alias{l2v}
\title{Convert List to Vector}
\usage{
l2v(x)
}
\arguments{
\item{x}{A list of numeric vectors.}
}
\value{
A numeric vector containing all values from all list elements,
concatenated in order.
}
\description{
Combines data stored in list mode into a single vector.
}
\details{
This utility function flattens a list structure into a single vector by
concatenating all elements sequentially.
}
\examples{
x <- list(c(1, 2, 3), c(4, 5), c(6, 7, 8, 9))
l2v(x)  # Returns c(1, 2, 3, 4, 5, 6, 7, 8, 9)

}
\keyword{utilities}
