% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{linWMWMC.sub2}
\alias{linWMWMC.sub2}
\title{Helper Function for Converting Contrasts to Probabilities}
\usage{
linWMWMC.sub2(L)
}
\arguments{
\item{L}{Vector of linear contrast values.}
}
\value{
Probability estimate P(L < 0) + 0.5 * P(L = 0).
}
\description{
Internal helper function that converts linear contrast values to
probability estimates for WMW-type statistics.
}
\details{
This function computes a probability estimate based on the proportion
of negative values plus half the proportion of zero values. Used
internally in WMW-based procedures.
}
\keyword{internal}
