% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{linWMWpb.lsub}
\alias{linWMWpb.lsub}
\title{Internal Helper for linWMWpb (Bootstrap Subsample)}
\usage{
linWMWpb.lsub(x, nreps = nreps, con = con, SEED = SEED)
}
\description{
Internal Helper for linWMWpb (Bootstrap Subsample)
}
\keyword{internal}
