% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{loc2difpb}
\alias{loc2difpb}
\title{Bootstrap Confidence Intervals for Location Difference Function}
\usage{
loc2difpb(x, y, est = median, alpha = 0.05, nboot = 2000, SEED = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix. For bivariate functions, the first variable.
For matrix functions, an n-by-p data matrix.}

\item{y}{Numeric vector. The second variable for bivariate correlations.
Can be \code{NULL} for some functions that accept a two-column matrix as \code{x}.}

\item{est}{Estimator function to use for computing location or other
statistics. Common choices include \code{mean}, \code{median}, \code{tmean} (trimmed mean),
\code{mom} (modified one-step M-estimator), \code{mest} (M-estimator).}
}
\value{
List with CI and p-value for the location of pairwise differences.
}
\description{
Bootstrap inference for the function relating location of X-Y to values of X.
}
\seealso{
\code{\link{loc2dif}}, \code{\link{shifthd}}
}
