% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{loc2plot}
\alias{loc2plot}
\title{Plot Two Location Estimates}
\usage{
loc2plot(x, y, plotfun = akerd, xlab = "X", ylab = "", ...)
}
\arguments{
\item{x}{Numeric vector or matrix. For bivariate functions, the first variable.
For matrix functions, an n-by-p data matrix.}

\item{y}{Numeric vector. The second variable for bivariate correlations.
Can be \code{NULL} for some functions that accept a two-column matrix as \code{x}.}

\item{plotfun}{Plotting function (default: \code{akerd}).}

\item{xlab, ylab}{Axis labels.}

\item{...}{Additional arguments passed to correlation or estimation functions.}
}
\value{
Plots (no return value).
}
\description{
Creates side-by-side plots of location estimates for two groups/variables.
}
\seealso{
\code{\link{loc2dif}}, \code{\link{akerd}}
}
