% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{lognormal.mean}
\alias{lognormal.mean}
\title{Mean of Lognormal Distribution}
\usage{
lognormal.mean(mu = 0, sig.sq = 1)
}
\arguments{
\item{mu}{Mean of the underlying normal distribution (default: 0).}

\item{sig.sq}{Variance of the underlying normal distribution (default: 1).}
}
\value{
Numeric value representing the mean: exp(mu + sig.sq/2).
}
\description{
Computes the theoretical mean of a lognormal distribution.
}
\seealso{
\code{\link{lognormal.mom}}
}
