% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lplotCIMCv2}
\alias{lplotCIMCv2}
\title{Bootstrap Helper for lplotcom2v2 (Parallel Processing)}
\usage{
lplotCIMCv2(data, x, y, pts1, pts2, npts, tr, span)
}
\arguments{
\item{data}{Bootstrap sample indices.}

\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{pts1}{Evaluation points for predictor 1.}

\item{pts2}{Evaluation points for predictor 2.}

\item{npts}{Number of evaluation points.}

\item{tr}{Trimming proportion for Winsorized SD.}

\item{span}{Span parameter for LOESS.}
}
\value{
Vector of length 2 containing mean.str1 and mean.str2.
}
\description{
Internal helper function for parallel bootstrap computation in \code{\link{lplotcomBCI}}.
Computes mean strength measures using \code{\link{lplotcom2v2}}.
}
\keyword{internal}
