% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{lsqtest4}
\alias{lsqtest4}
\title{Bootstrap Test for Single Regression Coefficient (HC4 Method)}
\usage{
lsqtest4(vstar, yhat, res, x, k)
}
\arguments{
\item{vstar}{Bootstrap resampled residuals.}

\item{yhat}{Fitted values from original regression.}

\item{res}{Residuals from original regression.}

\item{x}{Matrix of predictor variables.}

\item{k}{Integer. Index of the coefficient to test.}
}
\value{
Numeric. t-statistic for testing the k-th coefficient.
}
\description{
Internal function for testing a single regression coefficient using HC4
heteroscedasticity-robust standard errors within a bootstrap framework.
Computes a t-statistic for the k-th predictor.
}
\details{
This function is typically called repeatedly within a bootstrap loop. It:
\enumerate{
\item Creates bootstrap response: y* = yhat + res * vstar
\item Fits regression of y* on x
\item Computes HC4 variance-covariance matrix
\item Returns t-statistic for k-th coefficient
}
}
\seealso{
\code{\link{lsfitNci4}}, \code{\link{lstest4}}
}
\keyword{internal}
\keyword{regression}
