% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{lstest4}
\alias{lstest4}
\title{Bootstrap Test for All Regression Coefficients (HC4 Method)}
\usage{
lstest4(vstar, yhat, res, x)
}
\arguments{
\item{vstar}{Bootstrap resampled residuals.}

\item{yhat}{Fitted values from original regression.}

\item{res}{Residuals from original regression.}

\item{x}{Matrix of predictor variables.}
}
\value{
Numeric. Chi-square test statistic for testing all coefficients jointly.
}
\description{
Internal function for simultaneously testing all regression coefficients using
HC4 heteroscedasticity-robust covariance matrix within a bootstrap framework.
Computes a Wald-type chi-square statistic.
}
\details{
This function is typically called repeatedly within a bootstrap loop. It:
\enumerate{
\item Creates bootstrap response: y* = yhat + res * vstar
\item Fits regression of y* on x
\item Computes HC4 variance-covariance matrix
\item Returns Wald statistic: beta' Sigma^{-1} beta
}

The test statistic has an asymptotic chi-square distribution under the null
hypothesis that all slope coefficients are zero.
}
\seealso{
\code{\link{lsfitNci4}}, \code{\link{lsqtest4}}
}
\keyword{internal}
\keyword{regression}
