% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{mat2list}
\alias{mat2list}
\title{Convert Matrix to Grouped List (Exclude Grouping Column)}
\usage{
mat2list(m, grp.dat)
}
\arguments{
\item{m}{Data matrix or data frame containing variables to be grouped.}

\item{grp.dat}{Vector containing group identifiers (same length as \code{nrow(m)}).}
}
\value{
List where each element contains a matrix of data from \code{m} for one
group level (grouping column excluded).
}
\description{
Split matrix data into groups based on a grouping variable, excluding the
grouping column from the output. Similar to \code{\link{fac2list}} but for matrices.
}
\details{
This function divides rows of \code{m} into groups based on values in \code{grp.dat}.
Unlike \code{\link{mat2grp}}, the grouping variable is not included in the output
list elements.

Rows where \code{grp.dat} is \code{NA} are automatically removed.
}
\examples{
\dontrun{
# Divide columns 2-5 into groups based on group IDs in column 9
z <- mat2list(m[,2:5], m[,9])
# z[[1]] contains data from first group
# z[[2]] contains data from second group, etc.
}

}
\seealso{
\code{\link{mat2grp}}, \code{\link{fac2list}}, \code{\link{MAT2list}}
}
\keyword{manip}
