% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{matbin2v}
\alias{matbin2v}
\title{Extract Matrix Rows Based on Two-Variable Intervals}
\usage{
matbin2v(m, col = c(1, 2), int1 = c(0.5, 0.5), int2 = c(0.5, 0.5), INC = TRUE)
}
\arguments{
\item{m}{Matrix with two or more columns}

\item{col}{Vector of length 2 specifying which columns to use for filtering (default: c(1,2))}

\item{int1}{Numeric vector of length 2: interval \link{lower, upper} for first variable (default: c(.5,.5))}

\item{int2}{Numeric vector of length 2: interval \link{lower, upper} for second variable (default: c(.5,.5))}

\item{INC}{Logical; if TRUE uses inclusive bounds (>=, <=), if FALSE uses exclusive bounds (>, <). Default: TRUE}
}
\value{
Matrix containing only rows where values in specified columns fall within intervals
}
\description{
Extracts rows from a matrix where values in two specified columns fall within
specified intervals. Useful for subsetting multivariate data based on ranges.
}
\details{
Returns rows where:
\itemize{
\item m[,col\link{1}] is between int1\link{1} and int1\link{2}
\item m[,col\link{2}] is between int2\link{1} and int2\link{2}
}

Default intervals split at median (.5 quantile) for both variables.
}
\examples{
\dontrun{
# Return rows where column 1 is in [10,15] and column 3 is in [2,6]
matbin2v(m, c(1,3), c(10,15), c(2,6))
}

}
\keyword{manip}
