% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-core.R
\name{matl}
\alias{matl}
\alias{list2mat}
\title{Convert List to Matrix}
\usage{
matl(x)

list2mat(x)
}
\arguments{
\item{x}{A list where each element is a numeric vector.}
}
\value{
A matrix where the j-th column contains the j-th list element.
If list elements have different lengths, shorter vectors are padded with
NA values to match the longest vector.
}
\description{
Converts a list to matrix format, where each list element becomes a column
in the resulting matrix.
}
\details{
This is one of the most frequently called functions in the WRS package
(used by 219 functions). It's the inverse operation of \code{\link{listm}}.
When list elements have unequal lengths, the resulting matrix has as many
rows as the longest list element, with NA values filling shorter columns.
}
\examples{
# Convert list to matrix (equal lengths)
lst <- list(c(1, 2, 3), c(4, 5, 6), c(7, 8, 9))
matl(lst)

# Convert list with unequal lengths
lst2 <- list(c(1, 2), c(3, 4, 5, 6), c(7, 8, 9))
matl(lst2)

}
