% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{mmean}
\alias{mmean}
\title{Marginal Location Estimates}
\usage{
mmean(x, est = tmean, ...)
}
\arguments{
\item{x}{Matrix or list of numeric vectors.}

\item{est}{Location estimator function (default: \code{tmean}).}

\item{...}{Additional arguments passed to \code{est}.}
}
\value{
Vector or list of location estimates.
}
\description{
Computes location estimates for each column/element of a matrix or list.
}
\details{
Applies the specified estimator to each column (if matrix) or each element (if list).
}
\seealso{
\code{\link{tmean}}, \code{\link{lloc}}
}
