% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{olsJ2}
\alias{olsJ2}
\title{Test Equality of OLS Regression Parameters for Two Groups}
\usage{
olsJ2(
  x1,
  y1,
  x2,
  y2,
  xout = FALSE,
  outfun = outpro,
  STAND = TRUE,
  plotit = TRUE,
  xlab = "X",
  ylab = "Y",
  ISO = FALSE,
  ...
)
}
\arguments{
\item{x1}{Predictor(s) for group 1. Can be a vector or matrix.}

\item{y1}{Response variable for group 1.}

\item{x2}{Predictor(s) for group 2. Can be a vector or matrix.}

\item{y2}{Response variable for group 2.}

\item{xout}{Logical. If \code{TRUE}, eliminate leverage points. Default is \code{FALSE}.}

\item{outfun}{Outlier detection function. Default is \code{\link{outpro}}.}

\item{STAND}{Logical. Standardize data when detecting outliers. Default is \code{TRUE}.}

\item{plotit}{Logical. Create scatter plot with regression lines. Default is \code{TRUE}.}

\item{xlab}{X-axis label. Default is "X".}

\item{ylab}{Y-axis label. Default is "Y".}

\item{ISO}{Logical. If \code{TRUE}, test only slopes, ignoring intercepts. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to outlier detection function.}
}
\value{
A list from \code{\link{ols1way}} or \code{\link{ols1wayISO}} containing test results.
}
\description{
Tests the hypothesis that for two independent groups, all regression parameters
(or just slopes) are equal using OLS regression with Johansen-type test statistic.
}
\details{
Uses HC4 estimate of standard errors followed by Johansen-type test statistic.
When \code{ISO=TRUE}, only slopes are compared (intercepts can differ).
}
\seealso{
\code{\link{ols1way}}, \code{\link{ols1wayISO}}, \code{\link{ols2ci}}
}
\keyword{htest}
\keyword{regression}
