% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{ord.loc.fun}
\alias{ord.loc.fun}
\title{Objective Function for Ordinal Location Test Critical Values (Internal)}
\usage{
ord.loc.fun(a, init, iter, rem, Jm1, alpha)
}
\arguments{
\item{a}{Numeric scaling parameter being optimized.}

\item{init}{Numeric vector of initial critical value estimates.}

\item{iter}{Integer. Number of iterations/simulations.}

\item{rem}{Matrix of simulated p-values (iter × Jm1).}

\item{Jm1}{Integer. Number of groups minus 1 (J-1).}

\item{alpha}{Numeric. Target alpha level.}
}
\value{
Absolute difference between empirical and target alpha level.
}
\description{
Internal optimization objective function for determining critical p-values
in ordinal location tests. Used by \code{\link{ord.loc.crit.det}}.
}
\keyword{internal}
