% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outproMC.sub}
\alias{outproMC.sub}
\title{Helper Function for outproMC - Compute Projection Distances}
\usage{
outproMC.sub(B, Amat)
}
\arguments{
\item{B}{Numeric vector representing a direction for projection.}

\item{Amat}{Matrix of centered data points.}
}
\value{
Vector of projection distances.
}
\description{
Helper Function for outproMC - Compute Projection Distances
}
\keyword{internal}
