% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{pdep}
\alias{pdep}
\title{Estimate P(X < Y) for Dependent Variables}
\usage{
pdep(x, y, alpha = 0.05)
}
\arguments{
\item{x}{First numeric vector.}

\item{y}{Second numeric vector (same length as \code{x}).}

\item{alpha}{Significance level for confidence interval. Default is 0.05.}
}
\value{
List with components:
\item{phat}{Estimated probability P(X < Y)}
\item{ci}{Confidence interval for P(X < Y)}
}
\description{
For two dependent (paired) variables, estimates the probability that X is
less than Y with a confidence interval.
}
\details{
This function computes the proportion of pairs where X < Y and constructs
a binomial confidence interval via \code{\link{binomci}}. This provides
a simple nonparametric measure of the tendency for one variable to be
larger than the other.
}
\seealso{
\code{\link{cid}}, \code{\link{binomci}}, \code{\link{wmwloc}}
}
\keyword{htest}
\keyword{nonparametric}
