% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{perm.rho.sub}
\alias{perm.rho.sub}
\title{Helper Function for Permutation Test Statistic}
\usage{
perm.rho.sub(x, y)
}
\arguments{
\item{x}{First variable.}

\item{y}{Second variable.}
}
\value{
Standardized test statistic S = sqrt(n) * rho / tau, where
rho is the correlation and tau is an estimate of the standard deviation.
}
\description{
Internal helper function that computes the standardized test statistic
for the permutation test of Pearson's correlation.
}
\details{
This function computes a studentized version of Pearson's correlation
that is used in the DiCiccio-Romano permutation test. The statistic
accounts for the variance of the correlation coefficient.
}
\keyword{internal}
