% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{phiBY3}
\alias{phiBY3}
\title{Internal Function for BY3 Robust Logistic Regression}
\usage{
phiBY3(s, y, c3)
}
\arguments{
\item{s}{Numeric. Linear predictor value.}

\item{y}{Numeric. Binary response value (0 or 1).}

\item{c3}{Numeric. Tuning constant for the BY3 estimator.}
}
\value{
Numeric value of the phi function.
}
\description{
Computes the phi function for the Bianco-Yohai (BY3) robust logistic regression estimator.
This is an internal helper function used in the optimization process.
}
\details{
This function combines the rho function \code{\link{rhoBY3}} with first and second order
derivatives \code{\link{GBY3Fs}} and \code{\link{GBY3Fsm}}. Used internally for
robust logistic regression estimation.
}
\seealso{
\code{\link{rhoBY3}}, \code{\link{GBY3Fs}}, \code{\link{GBY3Fsm}},
\code{\link{wlogreg}}
}
\keyword{internal}
\keyword{regression}
