% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{PVALS}
\alias{PVALS}
\title{Compute Empirical CDF at Specified Values}
\usage{
PVALS(x, vals)
}
\arguments{
\item{x}{Numeric vector of data values}

\item{vals}{Numeric vector of values at which to evaluate the ECDF}
}
\value{
Numeric vector of the same length as \code{vals}, containing
the proportion of \code{x} values ≤ each corresponding value in \code{vals}
}
\description{
For each value in \code{vals}, computes the proportion of observations
in \code{x} that are less than or equal to that value. Essentially
evaluates the empirical cumulative distribution function (ECDF) at
specified points.
}
\details{
Missing values in \code{x} are automatically removed before computation.
Each output value is in \link{0, 1} representing the proportion of data points
at or below the corresponding threshold.
}
\examples{
x <- rnorm(100)
PVALS(x, c(-1, 0, 1))  # Proportions at -1, 0, 1
# Compare to theoretical normal CDF
pnorm(c(-1, 0, 1))
}
\seealso{
\code{\link[stats]{ecdf}} for creating empirical CDF functions
}
\keyword{distribution}
