% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{qrchkv2.sub2}
\alias{qrchkv2.sub2}
\title{Bootstrap Helper for Quantile Regression Lack-of-Fit Test}
\usage{
qrchkv2.sub2(xy, qval)
}
\arguments{
\item{xy}{Numeric matrix with response in last column, predictors in first columns}

\item{qval}{Numeric value specifying the quantile}
}
\value{
Numeric value: test statistic from \code{\link{qrchkv2}}
}
\description{
Internal helper function for \code{\link{qrchk}} that computes the test
statistic on simulated multivariate normal data for p-value estimation.
}
\seealso{
\code{\link{qrchk}}, \code{\link{qrchkv2}}
}
\keyword{internal}
