% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regts1}
\alias{regts1}
\title{Internal Bootstrap Helper for Regression Tests}
\usage{
regts1(vstar, yhat, res, mflag, x, tr)
}
\arguments{
\item{vstar}{Bootstrap resampling vector}

\item{yhat}{Vector of fitted values}

\item{res}{Vector of residuals}

\item{mflag}{Matrix of flags}

\item{x}{Predictor matrix}

\item{tr}{Trimming proportion}
}
\value{
Vector of resampled test statistics
}
\description{
Internal function used by \code{\link{regtest}} for computing bootstrap
test statistics. Not intended for direct use.
}
\keyword{internal}
