% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{ridgeGnullMC.sub}
\alias{ridgeGnullMC.sub}
\title{Helper Function for Ridge Regression Global Test}
\usage{
ridgeGnullMC.sub(x, p, regfun = regfun)
}
\arguments{
\item{x}{Matrix containing predictors and response (last column).}

\item{p}{Number of predictors.}

\item{regfun}{Regression function to use.}
}
\value{
F-test statistic value.
}
\description{
Internal helper function for computing F-test statistics in parallel
bootstrap procedures for ridge regression.
}
\details{
Used internally by \code{ridgeGnullMC} for parallel computation of bootstrap
F-test statistics in ridge regression global tests.
}
\keyword{internal}
