% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{RobRsq}
\alias{RobRsq}
\title{Robust R-Squared}
\usage{
RobRsq(x, y)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{Response vector}
}
\value{
Robust R-squared value
}
\description{
Computes robust R-squared measure using the robust package.
}
\keyword{regression}
\keyword{robust}
