% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{selvar.ind.ex}
\alias{selvar.ind.ex}
\title{Compare Group with Smallest Variance to All Others (Internal Helper)}
\usage{
selvar.ind.ex(x)
}
\arguments{
\item{x}{A matrix, data frame, or list of numeric vectors, one for each group.}
}
\value{
A numeric vector of p-values from comparing the group with smallest
variance to each of the other groups (length = J - 1, where J is the number
of groups).
}
\description{
Helper function for \code{\link{selvar.ind.MP}}. Identifies the group with
smallest variance and compares it to all other groups using the
\code{\link{varcom.IND.MP}} test.
}
\details{
This is an internal helper function used by \code{\link{selvar.ind.MP}} to
compute raw p-values. It:
\enumerate{
\item Computes variance for each group
\item Identifies the group with smallest variance
\item Compares that group to each other group using \code{\link{varcom.IND.MP}}
}

Users should typically use \code{\link{selvar.ind.MP}} instead, which provides
adjusted p-values and proper decision rules.
}
\seealso{
\code{\link{selvar.ind.MP}}, \code{\link{selvar.ind.crit}},
\code{\link{varcom.IND.MP}}
}
\keyword{internal}
