% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{sigmaBY3}
\alias{sigmaBY3}
\title{Sigma Function for BY3 Robust Regression (Internal)}
\usage{
sigmaBY3(sigma, s, y, c3)
}
\arguments{
\item{sigma}{Current estimate of scale parameter}

\item{s}{Residuals or standardized values}

\item{y}{Response variable values}

\item{c3}{Tuning constant for BY3 estimator}
}
\value{
Numeric value: mean of phi function evaluations
}
\description{
Internal function computing the sigma equation for the Bianco-Yohai (BY3)
robust regression estimator. Used in iterative scale estimation.
}
\details{
This is an internal function used by the Bianco-Yohai robust regression
procedure (BY3 variant). It computes the mean of the phi function applied
to standardized residuals, which is used in the iterative scale estimation
step of the algorithm.

The function is typically not called directly by users but is used internally
by robust regression functions.
}
\keyword{internal}
\keyword{regression}
