% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{sqmad}
\alias{sqmad}
\title{Squared Median Absolute Deviation}
\usage{
sqmad(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar: MAD(x)².
}
\description{
Computes the squared median absolute deviation (MAD²).
}
\details{
Returns MAD(x)², where MAD is the median absolute deviation. This squared
version can be used as a robust scale estimator in certain contexts where
a squared scale is needed.
}
\seealso{
\code{\link{mad}}
}
\keyword{robust}
\keyword{univar}
