% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample.R
\name{wmw.ancbse.sub}
\alias{wmw.ancbse.sub}
\title{Bootstrap Helper for WMW ANCOVA (Internal)}
\usage{
wmw.ancbse.sub(m, pts)
}
\arguments{
\item{m}{List containing bootstrapped data: x1, y1, x2, y2.}

\item{pts}{Matrix of covariate values.}
}
\value{
Vector of P(Y1 < Y2) estimates at each covariate value.
}
\description{
Internal helper function for \code{wmw.ancbse} that computes WMW ANCOVA estimates
for bootstrap samples.
}
\seealso{
\code{\link{wmw.ancbse}}, \code{\link{wmw.anc}}
}
\keyword{internal}
