% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils-core.R
\name{yuend}
\alias{yuend}
\title{Yuen's Test for Trimmed Means (Dependent Groups)}
\usage{
yuend(x, y, tr = 0.2, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector or matrix. For bivariate functions, the first variable.
For matrix functions, an n-by-p data matrix.}

\item{y}{Numeric vector. The second variable for bivariate correlations.
Can be \code{NULL} for some functions that accept a two-column matrix as \code{x}.}

\item{tr}{Trimming proportion (default: 0.2). Proportion of observations to
trim from each tail when computing winsorized correlation.}

\item{alpha}{Significance level for confidence intervals (default: 0.05).}
}
\value{
A list with components:
\item{test}{Test statistic}
\item{se}{Standard error of the difference}
\item{ci}{Confidence interval for the difference}
\item{p.value}{Two-sided p-value}
\item{dif}{Difference in trimmed means}
\item{n}{Number of pairs}
}
\description{
Performs Yuen's test for comparing trimmed means of two dependent (paired)
samples. This is a robust alternative to the paired t-test.
}
\details{
Pairs with missing values are eliminated. For methods that allow
missing values, use \code{rm2miss}.
}
\examples{
x <- rnorm(30, mean = 0)
y <- x + rnorm(30, mean = 0.5)
yuend(x, y)

}
\seealso{
\code{\link{yuen}} for independent groups
}
