% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{zdepth.sub}
\alias{zdepth.sub}
\title{Helper Function for zdepth - Compute Projection Outlyingness}
\usage{
zdepth.sub(x, theta, zloc = median, zscale = mad, pts = NA)
}
\description{
Helper Function for zdepth - Compute Projection Outlyingness
}
\keyword{internal}
