\name{WRS-package}
\Rdversion{1.1}
\alias{WRS-package}
\alias{WRS}
\docType{package}
\title{
Wilcox' Robust Statistics
}
\description{
This is a packaged version of the robust statistic functions provided by R.R. Wilcox (http://www-rcf.usc.edu/~rwilcox/). No help files are provided, please refer to the books for the usage of these functions.
}
\author{
Robust functions by R.R. Wilcox
Package maintainer: Felix Schönbrodt <felix@nicebread.de>
}
\references{
Wilcox, R. R. (2005). \emph{Introduction to robust estimation and hypothesis testing.} San Diego: Academic Press.

}

\keyword{ package }
